refactor(/stores):zustand 폴더명 변경
